pyproject.toml
[project]
name = "goretsky-integration-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "django-environ>=0.12.0",
    "django-import-export>=4.3.7",
    "djangorestframework>=3.16.0",
    "drf-standardized-errors>=0.14.1",
]
